backend/requirements.txt
flask==3.0.3
flask-sqlalchemy==3.1.1
flask-jwt-extended==4.6.0
flask-cors==4.0.0
flask-mail==0.10.0
flask-limiter==3.7.0
flask-migrate==4.0.7
psycopg2-binary==2.9.9
redis==5.0.7
openai>=1.52.0
python-dotenv==1.0.1
bcrypt==4.1.3
Werkzeug==3.0.3
marshmallow==3.21.3
fpdf2==2.7.9
Pillow==10.3.0
celery==5.4.0
kombu==5.3.7
gunicorn==22.0.0
